New ping route, minor fixes.
Former-commit-id: 42e0e5c88383a055fdea74a10d56b5cb0c9c77e2

+++ client/src/main/scala/JsonSerializers.scala
@@ -12,7 +12,7 @@ object JsonSerializers {
     case None => Js.Null
   }
   private implicit val intOptionWriter = Writer[Option[Int]] {
-    case Some(i) => Js.Num(i)
+    case Some(i) => Js.Num(i.toDouble)
     case None => Js.Null
   }
   private implicit val boolOptionWriter = Writer[Option[Boolean]] {
