+++ shared/src/main/scala/models/rules/GameRulesHelper.scala
@@ -10,7 +10,7 @@ import scala.util.Random
 
 trait GameRulesHelper { this: GameRules =>
   def newGame(gameId: UUID, seed: Int, withRules: String) = {
-    val rng = new Random(new java.util.Random(seed))
+    val rng = new Random(new java.util.Random(seed.toLong))
     val maxPlayers = 1
     val lowRank = if (deckOptions.lowRank == Rank.Unknown) {
       deckOptions.ranks(rng.nextInt(Rank.all.length))
